Added more go versions to build matrix
This change adds additional builds to the build matrix.

Go version 1.2 and 1.3 are still supported by Ubuntu 14.04 LTS.
So, even though go users tend to use the latest stable it would
be nice to know at a glance that this library is passing for legacy
users.

Go tip has also been added and placed under allowed failures for
bleeding edge users as well.