fosite.Request: Update Request ID in Merge(request)
Updating Request.Merge() to also merge the request ID from the passed in
request, since in all cases where Merge() is called, the receiver object
has an uninitialized RequestID value. This fixes an issue where
introspectAccessToken() is returning the wrong request ID, when it
shoudl be returning the value returned by the
storage.AccessTokenSession().