fix: empty client secret via basic auth header means "none" authn (#655)
The existing client authentication code treats an empty client_secret
query parameter to be equivalent to "none" authentication instead of
"client_secret_post."

This change updates the basic auth check to be consistent with this.
That is, an empty secret via the basic auth header is considered to
mean "none" instead of "client_secret_basic."

The "golang.org/x/oauth2" library probes for both methods of
authentication, starting with the basic auth header approach first.

As required, both client ID and secret are encoded in one header:

https://github.com/golang/oauth2/blob/ee480838109b20d468babcb00b7027c82f962065/internal/token.go#L174-L176

Since this code path is taken regardless of if the client is public
or confidential, without this change, it results in an error in the
fosite code.  Since the request fails, the library code simply tries
with the other approach, which works because the client secret is
empty.  Needless to say, it is inefficient to force public clients
to make multiple requests.

Signed-off-by: Monis Khan <mok@vmware.com>